[branches/unstable] -프로젝트 문서 추가 -DB 관련 코드 정리 중 -몹 생성 버튼하고 관련해서 핸들러 사용 없애고 다른 방식으로 구현함.  몹 생성 버튼만 직접 그리는 방식으로 했음.  터치 이벤트 구조도 살짝 변경됨.
-타이머 3배속 때 생기는 오차 해결
-어플리케이션이 가려져서 pause 되거나 destroy 된 후 다시 화면 위로 올라올 때 발생하던 dialog 에러 해결
-기기 잠금화면으로 넘어가서 화면 회전시 발생하는 카메라 버그 해결

+++ branches/unstable/src/hjsi/common/Camera.java
@@ -114,7 +114,7 @@ public class Camera {
     worldMarginSpace = margin;
   }
 
-  public void setViewportSize(int width, int height) {
+  public void setViewportSize(float width, float height) {
     if (viewport == null) {
       viewport = new RectF(0, 0, width, height);
     } else {
